Change deltaTime to float

+++ LOLIS2D/inc/Time.hpp
@@ -5,6 +5,6 @@ namespace LOLIS2D
 	class Time final
 	{
 	public:
-		static double deltaTime;
+		static float deltaTime;
 	};
 }
